requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
autogen
openai>=1.0.0
